+++ tests/__init__.py
@@ -12,28 +12,8 @@ from templatest import BaseTemplate as _BaseTemplate
 from templatest import templates as _templates
 from templatest.utils import VarPrefix as _VarPrefix
 
-from docsig.messages import (
-    E101,
-    E102,
-    E103,
-    E104,
-    E105,
-    E106,
-    E107,
-    E109,
-    E110,
-    E111,
-    E112,
-    E113,
-    E114,
-    E115,
-    E116,
-    E201,
-    E202,
-    E203,
-    E204,
-    E,
-)
+from docsig.messages import TEMPLATE as T
+from docsig.messages import E
 
 MockMainType = t.Callable[..., int]
 InitFileFixtureType = t.Callable[[str], Path]
@@ -145,7 +125,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E113
+        return E[113].fstring(T)
 
 
 @_templates.register
@@ -300,7 +280,7 @@ def function(param1, param2, param3):
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -398,7 +378,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E101
+        return E[101].fstring(T)
 
 
 @_templates.register
@@ -441,7 +421,7 @@ def function(param1, param2) -> None:
 
     @property
     def expected(self) -> str:
-        return E102
+        return E[102].fstring(T)
 
 
 @_templates.register
@@ -459,7 +439,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E103
+        return E[103].fstring(T)
 
 
 @_templates.register
@@ -479,7 +459,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E104
+        return E[104].fstring(T)
 
 
 @_templates.register
@@ -499,7 +479,7 @@ def function(param1, param2, param3) -> int:
 
     @property
     def expected(self) -> str:
-        return E105
+        return E[105].fstring(T)
 
 
 @_templates.register
@@ -519,7 +499,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E106
+        return E[106].fstring(T)
 
 
 @_templates.register
@@ -539,7 +519,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E107
+        return E[107].fstring(T)
 
 
 @_templates.register
@@ -672,7 +652,7 @@ def function_1({CROSS}param1, {CROSS}param2, {CROSS}param3) -> {CHECK}None:
     :param param1: {CROSS}
     \"\"\"
 
-{E101}
+{E[101].fstring(T)}
 
 module/file.py:11
 -----------------
@@ -683,7 +663,7 @@ def function_2({CHECK}param1, {CHECK}param2, {CROSS}None) -> {CHECK}None:
     :param param3: {CROSS}
     \"\"\"
 
-{E102}
+{E[102].fstring(T)}
 
 module/file.py:19
 -----------------
@@ -694,7 +674,7 @@ def function_3({CHECK}param1, {CHECK}param2, {CROSS}param3) -> {CHECK}None:
     :param None: {CROSS}
     \"\"\"
 
-{E103}
+{E[103].fstring(T)}
 
 """
 
@@ -760,7 +740,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E105
+        return E[105].fstring(T)
 
 
 @_templates.register
@@ -896,7 +876,7 @@ def get_post(
 
     @property
     def expected(self) -> str:
-        return E116
+        return E[116].fstring(T)
 
 
 @_templates.register
@@ -913,7 +893,7 @@ def function(param1) -> None:
 
     @property
     def expected(self) -> str:
-        return E115
+        return E[115].fstring(T)
 
 
 @_templates.register
@@ -1008,7 +988,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E105
+        return E[105].fstring(T)
 
 
 @_templates.register
@@ -1082,7 +1062,7 @@ class MutableSet(_t.MutableSet[T]):
 
     @property
     def expected(self) -> str:
-        return E113
+        return E[113].fstring(T)
 
 
 @_templates.register
@@ -1096,7 +1076,7 @@ def function() -> int:
 
     @property
     def expected(self) -> str:
-        return E113
+        return E[113].fstring(T)
 
 
 @_templates.register
@@ -1133,7 +1113,7 @@ def function():
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -1147,7 +1127,7 @@ def function():
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -1165,7 +1145,7 @@ def function(arg, param2) -> None:
 
     @property
     def expected(self) -> str:
-        return E110
+        return E[110].fstring(T)
 
 
 @_templates.register
@@ -1258,7 +1238,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -1354,7 +1334,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E111
+        return E[111].fstring(T)
 
 
 @_templates.register
@@ -1399,7 +1379,7 @@ def function(self) -> int:
 
     @property
     def expected(self) -> str:
-        return E105
+        return E[105].fstring(T)
 
 
 @_templates.register
@@ -1437,7 +1417,7 @@ def method(self):
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -1451,7 +1431,7 @@ def __init__(param1, param2) -> None:
 
     @property
     def expected(self) -> str:
-        return E113
+        return E[113].fstring(T)
 
 
 @_templates.register
@@ -1543,7 +1523,7 @@ class MutableSet:
 
     @property
     def expected(self) -> str:
-        return E113
+        return E[113].fstring(T)
 
 
 @_templates.register
@@ -1587,7 +1567,7 @@ def function(param) -> None:
 
     @property
     def expected(self) -> str:
-        return E112
+        return E[112].fstring(T)
 
 
 @_templates.register
@@ -1685,7 +1665,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E113
+        return E[113].fstring(T)
 
 
 @_templates.register
@@ -1877,7 +1857,7 @@ def function(param1, param2, param3):
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -1995,7 +1975,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E101
+        return E[101].fstring(T)
 
 
 @_templates.register
@@ -2019,7 +1999,7 @@ def function(param1, param2) -> None:
 
     @property
     def expected(self) -> str:
-        return E102
+        return E[102].fstring(T)
 
 
 @_templates.register
@@ -2041,7 +2021,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E103
+        return E[103].fstring(T)
 
 
 @_templates.register
@@ -2070,7 +2050,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E104
+        return E[104].fstring(T)
 
 
 @_templates.register
@@ -2095,7 +2075,7 @@ def function(param1, param2, param3) -> int:
 
     @property
     def expected(self) -> str:
-        return E105
+        return E[105].fstring(T)
 
 
 @_templates.register
@@ -2121,7 +2101,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E106
+        return E[106].fstring(T)
 
 
 @_templates.register
@@ -2286,7 +2266,7 @@ def function_1({CROSS}param1, {CROSS}param2, {CROSS}param3) -> {CHECK}None:
     :param param1: {CROSS}
     \"\"\"
 
-{E101}
+{E[101].fstring(T)}
 
 module/file.py:16
 -----------------
@@ -2297,7 +2277,7 @@ def function_2({CHECK}param1, {CHECK}param2, {CROSS}None) -> {CHECK}None:
     :param param3: {CROSS}
     \"\"\"
 
-{E102}
+{E[102].fstring(T)}
 
 module/file.py:29
 -----------------
@@ -2308,7 +2288,7 @@ def function_3({CHECK}param1, {CHECK}param2, {CROSS}param3) -> {CHECK}None:
     :param None: {CROSS}
     \"\"\"
 
-{E103}
+{E[103].fstring(T)}
 
 """
 
@@ -2381,7 +2361,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E105
+        return E[105].fstring(T)
 
 
 @_templates.register
@@ -2425,7 +2405,7 @@ def function(param1, **kwargs) -> None:
 
     @property
     def expected(self) -> str:
-        return E103
+        return E[103].fstring(T)
 
 
 @_templates.register
@@ -2447,7 +2427,7 @@ def function(param1, **kwargs) -> None:
 
     @property
     def expected(self) -> str:
-        return E101
+        return E[101].fstring(T)
 
 
 @_templates.register
@@ -2643,7 +2623,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E105
+        return E[105].fstring(T)
 
 
 @_templates.register
@@ -2704,7 +2684,7 @@ class MutableSet(_t.MutableSet[T]):
 
     @property
     def expected(self) -> str:
-        return E113
+        return E[113].fstring(T)
 
 
 @_templates.register
@@ -2718,7 +2698,7 @@ def function() -> int:
 
     @property
     def expected(self) -> str:
-        return E113
+        return E[113].fstring(T)
 
 
 @_templates.register
@@ -2738,7 +2718,7 @@ def function():
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -2752,7 +2732,7 @@ def function():
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -2774,7 +2754,7 @@ def function(arg, param2) -> None:
 
     @property
     def expected(self) -> str:
-        return E110
+        return E[110].fstring(T)
 
 
 @_templates.register
@@ -2885,7 +2865,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -3005,7 +2985,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E111
+        return E[111].fstring(T)
 
 
 @_templates.register
@@ -3058,7 +3038,7 @@ def function(self) -> int:
 
     @property
     def expected(self) -> str:
-        return E105
+        return E[105].fstring(T)
 
 
 @_templates.register
@@ -3102,7 +3082,7 @@ def method(self):
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -3116,7 +3096,7 @@ def __init__(param1, param2) -> None:
 
     @property
     def expected(self) -> str:
-        return E113
+        return E[113].fstring(T)
 
 
 @_templates.register
@@ -3216,7 +3196,7 @@ class MutableSet:
 
     @property
     def expected(self) -> str:
-        return E113
+        return E[113].fstring(T)
 
 
 @_templates.register
@@ -3268,7 +3248,7 @@ def function(param) -> None:
 
     @property
     def expected(self) -> str:
-        return E112
+        return E[112].fstring(T)
 
 
 @_templates.register
@@ -3648,7 +3628,7 @@ def function(param1, param2, param3):
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -3766,7 +3746,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E101
+        return E[101].fstring(T)
 
 
 @_templates.register
@@ -3790,7 +3770,7 @@ def function(param1, param2) -> None:
 
     @property
     def expected(self) -> str:
-        return E102
+        return E[102].fstring(T)
 
 
 @_templates.register
@@ -3812,7 +3792,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E103
+        return E[103].fstring(T)
 
 
 @_templates.register
@@ -3841,7 +3821,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E104
+        return E[104].fstring(T)
 
 
 @_templates.register
@@ -3866,7 +3846,7 @@ def function(param1, param2, param3) -> int:
 
     @property
     def expected(self) -> str:
-        return E105
+        return E[105].fstring(T)
 
 
 @_templates.register
@@ -3892,7 +3872,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E106
+        return E[106].fstring(T)
 
 
 @_templates.register
@@ -4057,7 +4037,7 @@ def function_1({CROSS}param1, {CROSS}param2, {CROSS}param3) -> {CHECK}None:
     :param param1: {CROSS}
     \"\"\"
 
-{E101}
+{E[101].fstring(T)}
 
 module/file.py:16
 -----------------
@@ -4068,7 +4048,7 @@ def function_2({CHECK}param1, {CHECK}param2, {CROSS}None) -> {CHECK}None:
     :param param3: {CROSS}
     \"\"\"
 
-{E102}
+{E[102].fstring(T)}
 
 module/file.py:29
 -----------------
@@ -4079,7 +4059,7 @@ def function_3({CHECK}param1, {CHECK}param2, {CROSS}param3) -> {CHECK}None:
     :param None: {CROSS}
     \"\"\"
 
-{E103}
+{E[103].fstring(T)}
 
 """
 
@@ -4179,7 +4159,7 @@ def function(param1, **kwargs) -> None:
 
     @property
     def expected(self) -> str:
-        return E103
+        return E[103].fstring(T)
 
 
 @_templates.register
@@ -4201,7 +4181,7 @@ def function(param1, **kwargs) -> None:
 
     @property
     def expected(self) -> str:
-        return E101
+        return E[101].fstring(T)
 
 
 @_templates.register
@@ -4410,7 +4390,7 @@ def function():
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -4432,7 +4412,7 @@ def function(arg, param2) -> None:
 
     @property
     def expected(self) -> str:
-        return E110
+        return E[110].fstring(T)
 
 
 @_templates.register
@@ -4621,7 +4601,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E111
+        return E[111].fstring(T)
 
 
 @_templates.register
@@ -4674,7 +4654,7 @@ def function(self) -> int:
 
     @property
     def expected(self) -> str:
-        return E105
+        return E[105].fstring(T)
 
 
 @_templates.register
@@ -4718,7 +4698,7 @@ def method(self):
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -4825,7 +4805,7 @@ def function(param) -> None:
 
     @property
     def expected(self) -> str:
-        return E112
+        return E[112].fstring(T)
 
 
 @_templates.register
@@ -4922,7 +4902,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E104
+        return E[104].fstring(T)
 
 
 @_templates.register
@@ -4986,7 +4966,7 @@ def get_post(
 
     @property
     def expected(self) -> str:
-        return E116
+        return E[116].fstring(T)
 
 
 @_templates.register
@@ -5106,7 +5086,7 @@ def function():
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -5183,7 +5163,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E111
+        return E[111].fstring(T)
 
 
 @_templates.register
@@ -5407,7 +5387,7 @@ def function(param1, param2, param3):
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -5511,7 +5491,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E101
+        return E[101].fstring(T)
 
 
 @_templates.register
@@ -5531,7 +5511,7 @@ def function(param1, param2) -> None:
 
     @property
     def expected(self) -> str:
-        return E102
+        return E[102].fstring(T)
 
 
 @_templates.register
@@ -5550,7 +5530,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E103
+        return E[103].fstring(T)
 
 
 @_templates.register
@@ -5573,7 +5553,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E104
+        return E[104].fstring(T)
 
 
 @_templates.register
@@ -5594,7 +5574,7 @@ def function(param1, param2, param3) -> int:
 
     @property
     def expected(self) -> str:
-        return E105
+        return E[105].fstring(T)
 
 
 @_templates.register
@@ -5615,7 +5595,7 @@ def function(param1, param2, param3) -> None:
 
     @property
     def expected(self) -> str:
-        return E106
+        return E[106].fstring(T)
 
 
 @_templates.register
@@ -5755,7 +5735,7 @@ def function_1({CROSS}param1, {CROSS}param2, {CROSS}param3) -> {CHECK}None:
     :param param1: {CROSS}
     \"\"\"
 
-{E101}
+{E[101].fstring(T)}
 
 module/file.py:12
 -----------------
@@ -5766,7 +5746,7 @@ def function_2({CHECK}param1, {CHECK}param2, {CROSS}None) -> {CHECK}None:
     :param param3: {CROSS}
     \"\"\"
 
-{E102}
+{E[102].fstring(T)}
 
 module/file.py:21
 -----------------
@@ -5777,7 +5757,7 @@ def function_3({CHECK}param1, {CHECK}param2, {CROSS}param3) -> {CHECK}None:
     :param None: {CROSS}
     \"\"\"
 
-{E103}
+{E[103].fstring(T)}
 
 """
 
@@ -5866,7 +5846,7 @@ def function(param1, **kwargs) -> None:
 
     @property
     def expected(self) -> str:
-        return E103
+        return E[103].fstring(T)
 
 
 @_templates.register
@@ -5885,7 +5865,7 @@ def function(param1, **kwargs) -> None:
 
     @property
     def expected(self) -> str:
-        return E101
+        return E[101].fstring(T)
 
 
 @_templates.register
@@ -6067,7 +6047,7 @@ def function():
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -6086,7 +6066,7 @@ def function(arg, param2) -> None:
 
     @property
     def expected(self) -> str:
-        return E110
+        return E[110].fstring(T)
 
 
 @_templates.register
@@ -6249,7 +6229,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E111
+        return E[111].fstring(T)
 
 
 @_templates.register
@@ -6296,7 +6276,7 @@ def function(self) -> int:
 
     @property
     def expected(self) -> str:
-        return E105
+        return E[105].fstring(T)
 
 
 @_templates.register
@@ -6336,7 +6316,7 @@ def method(self):
 
     @property
     def expected(self) -> str:
-        return E109
+        return E[109].fstring(T)
 
 
 @_templates.register
@@ -6431,7 +6411,7 @@ def function(param) -> None:
 
     @property
     def expected(self) -> str:
-        return E112
+        return E[112].fstring(T)
 
 
 @_templates.register
@@ -6498,7 +6478,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E114
+        return E[114].fstring(T)
 
 
 @_templates.register
@@ -6539,7 +6519,7 @@ def check_stuff({CHECK}str_lin, {CHECK}a) -> {CHECK}bool:
     :return: {CHECK}
     \"\"\"
 
-{E115}
+{E[115].fstring(T)}
 """
 
 
@@ -6623,7 +6603,7 @@ class Params(_MutableSequence[Param]):
 
     @property
     def expected(self) -> str:
-        return E113
+        return E[113].fstring(T)
 
 
 @_templates.register
@@ -6661,7 +6641,7 @@ def example(some_input: int) -> int:
 
     @property
     def expected(self) -> str:
-        return E103
+        return E[103].fstring(T)
 
 
 @_templates.register
@@ -6839,7 +6819,7 @@ def process({CHECK}response) -> {CROSS}str:
     :return: {CROSS}
     \"\"\"
 
-{E105}
+{E[105].fstring(T)}
 
 """
 
@@ -6885,7 +6865,7 @@ def process({CROSS}response) -> {CHECK}str:
     :return: {CHECK}
     \"\"\"
 
-{E103}
+{E[103].fstring(T)}
 
 """
 
@@ -7025,7 +7005,7 @@ def process({CHECK}response) -> {CROSS}None:
     :return: {CROSS}
     \"\"\"
 
-{E104}
+{E[104].fstring(T)}
 
 """
 
@@ -7101,7 +7081,7 @@ def process({CHECK}response) -> {CROSS}str:
     :return: {CROSS}
     \"\"\"
 
-{E105}
+{E[105].fstring(T)}
 
 """
 
@@ -7144,7 +7124,7 @@ def process({CROSS}response) -> {CHECK}str:
     :return: {CHECK}
     \"\"\"
 
-{E103}
+{E[103].fstring(T)}
 
 """
 
@@ -7272,7 +7252,7 @@ def process({CHECK}response) -> {CROSS}None:
     :return: {CROSS}
     \"\"\"
 
-{E104}
+{E[104].fstring(T)}
 
 """
 
@@ -7358,7 +7338,7 @@ def function_2({CHECK}param1, {CHECK}param2, {CROSS}None) -> {CHECK}None:
     :param param3: {CROSS}
     \"\"\"
 
-{E102}
+{E[102].fstring(T)}
 
 module/file.py:19
 -----------------
@@ -7369,7 +7349,7 @@ def function_3({CHECK}param1, {CHECK}param2, {CROSS}param3) -> {CHECK}None:
     :param None: {CROSS}
     \"\"\"
 
-{E103}
+{E[103].fstring(T)}
 
 """
 
@@ -7453,7 +7433,7 @@ def function_1({CROSS}param1, {CROSS}param2, {CROSS}param3) -> {CHECK}None:
     :param param1: {CROSS}
     \"\"\"
 
-{E101}
+{E[101].fstring(T)}
 
 """
 
@@ -7501,7 +7481,7 @@ def function_1({CROSS}param1, {CROSS}param2, {CROSS}param3) -> {CHECK}None:
     :param param1: {CROSS}
     \"\"\"
 
-{E101}
+{E[101].fstring(T)}
 
 module/file.py:11
 -----------------
@@ -7512,7 +7492,7 @@ def function_2({CHECK}param1, {CHECK}param2, {CROSS}None) -> {CHECK}None:
     :param param3: {CROSS}
     \"\"\"
 
-{E102}
+{E[102].fstring(T)}
 
 """
 
@@ -7561,7 +7541,7 @@ def function_1({CROSS}param1, {CROSS}param2, {CROSS}param3) -> {CHECK}None:
     :param param1: {CROSS}
     \"\"\"
 
-{E101}
+{E[101].fstring(T)}
 
 module/file.py:21
 -----------------
@@ -7572,7 +7552,7 @@ def function_3({CHECK}param1, {CHECK}param2, {CROSS}param3) -> {CHECK}None:
     :param None: {CROSS}
     \"\"\"
 
-{E103}
+{E[103].fstring(T)}
 
 """
 
@@ -7647,7 +7627,7 @@ def function_1({CROSS}param1, {CROSS}param2, {CROSS}param3) -> {CHECK}None:
     :param param1: {CROSS}
     \"\"\"
 
-{E101}
+{E[101].fstring(T)}
 
 module/file.py:21
 -----------------
@@ -7658,7 +7638,7 @@ def function_3({CHECK}param1, {CHECK}param2, {CROSS}param3) -> {CHECK}None:
     :param None: {CROSS}
     \"\"\"
 
-{E103}
+{E[103].fstring(T)}
 
 module/file.py:47
 -----------------
@@ -7669,7 +7649,7 @@ def function_6({CHECK}param1, {CHECK}param2, {CROSS}param3) -> {CHECK}None:
     :param None: {CROSS}
     \"\"\"
 
-{E103}
+{E[103].fstring(T)}
 
 """
 
@@ -7755,7 +7735,7 @@ def function_2({CHECK}param1, {CHECK}param2, {CROSS}None) -> {CHECK}None:
     :param param3: {CROSS}
     \"\"\"
 
-{E102}
+{E[102].fstring(T)}
 
 """
 
@@ -7795,7 +7775,7 @@ def function_1({CROSS}param1, {CROSS}param2, {CROSS}param3) -> {CHECK}None:
     :param param1: {CROSS}
     \"\"\"
 
-{E202.format(directive="ena")}
+{E[202].fstring(T).format(directive="ena")}
 
 module/file.py:11
 -----------------
@@ -7806,7 +7786,7 @@ def function_2({CHECK}param1, {CHECK}param2, {CROSS}None) -> {CHECK}None:
     :param param3: {CROSS}
     \"\"\"
 
-{E202.format(directive="ena")}
+{E[202].fstring(T).format(directive="ena")}
 
 """
 
@@ -7846,8 +7826,8 @@ def function_1({CHECK}param1, {CHECK}param2, {CROSS}None) -> {CHECK}None:
     :param param3: {CROSS}
     \"\"\"
 
-{E102}
-{E201.format(directive="disa")}
+{E[102].fstring(T)}
+{E[201].fstring(T).format(directive="disa")}
 
 module/file.py:11
 -----------------
@@ -7858,8 +7838,8 @@ def function_2({CROSS}param1, {CROSS}param2, {CROSS}param3) -> {CHECK}None:
     :param param1: {CROSS}
     \"\"\"
 
-{E101}
-{E201.format(directive="disa")}
+{E[101].fstring(T)}
+{E[201].fstring(T).format(directive="disa")}
 
 """
 
@@ -7951,8 +7931,8 @@ def function_1({CROSS}param1, {CROSS}param2, {CROSS}param3) -> {CHECK}None:
     :param param1: {CROSS}
     \"\"\"
 
-{E101}
-{E201.format(directive="unknown")}
+{E[101].fstring(T)}
+{E[201].fstring(T).format(directive="unknown")}
 
 module/file.py:14
 -----------------
@@ -7963,8 +7943,8 @@ def function_2({CHECK}param1, {CHECK}param2, {CROSS}None) -> {CHECK}None:
     :param param3: {CROSS}
     \"\"\"
 
-{E102}
-{E201.format(directive="unknown")}
+{E[102].fstring(T)}
+{E[201].fstring(T).format(directive="unknown")}
 
 module/file.py:23
 -----------------
@@ -7975,10 +7955,10 @@ def function_3({CHECK}param1, {CHECK}param2, {CROSS}param3) -> {CHECK}None:
     :param None: {CROSS}
     \"\"\"
 
-{E103}
-{E201.format(directive="unknown")}
-{E204.format(directive="enable", option="unknown")}
-{E204.format(directive="enable", option="unknown-the-third")}
+{E[103].fstring(T)}
+{E[201].fstring(T).format(directive="unknown")}
+{E[204].fstring(T).format(directive="enable", option="unknown")}
+{E[204].fstring(T).format(directive="enable", option="unknown-the-third")}
 
 module/file.py:33
 -----------------
@@ -7990,8 +7970,8 @@ def function_4({CHECK}param1, {CHECK}param2, {CHECK}param3) -> {CROSS}None:
     :return: {CROSS}
     \"\"\"
 
-{E104}
-{E201.format(directive="unknown")}
+{E[104].fstring(T)}
+{E[201].fstring(T).format(directive="unknown")}
 
 module/file.py:43
 -----------------
@@ -8003,8 +7983,8 @@ def function_5({CHECK}param1, {CHECK}param2, {CHECK}param3) -> {CROSS}int:
     :return: {CROSS}
     \"\"\"
 
-{E105}
-{E201.format(directive="unknown")}
+{E[105].fstring(T)}
+{E[201].fstring(T).format(directive="unknown")}
 
 module/file.py:52
 -----------------
@@ -8016,10 +7996,10 @@ def f6({CHECK}param, {CROSS}param2, {CROSS}param3, {CROSS}None) -> {CHECK}None:
     :param param3: {CROSS}
     \"\"\"
 
-{E101}
-{E102}
-{E106}
-{E201.format(directive="unknown")}
+{E[101].fstring(T)}
+{E[102].fstring(T)}
+{E[106].fstring(T)}
+{E[201].fstring(T).format(directive="unknown")}
 
 module/file.py:62
 -----------------
@@ -8031,11 +8011,11 @@ def f7({CHECK}param, {CROSS}param2, {CROSS}param3, {CROSS}None) -> {CHECK}None:
     :param None: {CROSS}
     \"\"\"
 
-{E101}
-{E102}
-{E106}
-{E107}
-{E201.format(directive="unknown")}
+{E[101].fstring(T)}
+{E[102].fstring(T)}
+{E[106].fstring(T)}
+{E[107].fstring(T)}
+{E[201].fstring(T).format(directive="unknown")}
 
 """
 
@@ -8127,8 +8107,8 @@ def function_1({CROSS}param1, {CROSS}param2, {CROSS}param3) -> {CHECK}None:
     :param param1: {CROSS}
     \"\"\"
 
-{E101}
-{E201.format(directive="unknown")}
+{E[101].fstring(T)}
+{E[201].fstring(T).format(directive="unknown")}
 
 module/file.py:14
 -----------------
@@ -8139,8 +8119,8 @@ def function_2({CHECK}param1, {CHECK}param2, {CROSS}None) -> {CHECK}None:
     :param param3: {CROSS}
     \"\"\"
 
-{E102}
-{E201.format(directive="unknown")}
+{E[102].fstring(T)}
+{E[201].fstring(T).format(directive="unknown")}
 
 module/file.py:23
 -----------------
@@ -8151,10 +8131,10 @@ def function_3({CHECK}param1, {CHECK}param2, {CROSS}param3) -> {CHECK}None:
     :param None: {CROSS}
     \"\"\"
 
-{E103}
-{E201.format(directive="unknown")}
-{E204.format(directive="enable", option="unknown")}
-{E204.format(directive="enable", option="unknown-the-third")}
+{E[103].fstring(T)}
+{E[201].fstring(T).format(directive="unknown")}
+{E[204].fstring(T).format(directive="enable", option="unknown")}
+{E[204].fstring(T).format(directive="enable", option="unknown-the-third")}
 
 module/file.py:33
 -----------------
@@ -8166,8 +8146,8 @@ def function_4({CHECK}param1, {CHECK}param2, {CHECK}param3) -> {CROSS}None:
     :return: {CROSS}
     \"\"\"
 
-{E104}
-{E201.format(directive="unknown")}
+{E[104].fstring(T)}
+{E[201].fstring(T).format(directive="unknown")}
 
 module/file.py:43
 -----------------
@@ -8179,8 +8159,8 @@ def function_5({CHECK}param1, {CHECK}param2, {CHECK}param3) -> {CROSS}int:
     :return: {CROSS}
     \"\"\"
 
-{E105}
-{E201.format(directive="unknown")}
+{E[105].fstring(T)}
+{E[201].fstring(T).format(directive="unknown")}
 
 module/file.py:52
 -----------------
@@ -8192,10 +8172,10 @@ def f6({CHECK}param, {CROSS}param2, {CROSS}param3, {CROSS}None) -> {CHECK}None:
     :param param3: {CROSS}
     \"\"\"
 
-{E101}
-{E102}
-{E106}
-{E201.format(directive="unknown")}
+{E[101].fstring(T)}
+{E[102].fstring(T)}
+{E[106].fstring(T)}
+{E[201].fstring(T).format(directive="unknown")}
 
 module/file.py:62
 -----------------
@@ -8207,11 +8187,11 @@ def f7({CHECK}param, {CROSS}param2, {CROSS}param3, {CROSS}None) -> {CHECK}None:
     :param None: {CROSS}
     \"\"\"
 
-{E101}
-{E102}
-{E106}
-{E107}
-{E201.format(directive="unknown")}
+{E[101].fstring(T)}
+{E[102].fstring(T)}
+{E[106].fstring(T)}
+{E[107].fstring(T)}
+{E[201].fstring(T).format(directive="unknown")}
 
 """
 
@@ -8243,8 +8223,8 @@ def function_3({CHECK}param1, {CHECK}param2, {CROSS}param3) -> {CHECK}None:
     :param None: {CROSS}
     \"\"\"
 
-{E103}
-{E204.format(directive="enable", option="unknown")}
+{E[103].fstring(T)}
+{E[204].fstring(T).format(directive="enable", option="unknown")}
 
 """
 
@@ -8269,7 +8249,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E103
+        return E[103].fstring(T)
 
 
 @_templates.register
@@ -8291,7 +8271,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E103
+        return E[103].fstring(T)
 
 
 @_templates.register
@@ -8314,7 +8294,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E103
+        return E[103].fstring(T)
 
 
 @_templates.register
@@ -8337,7 +8317,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E103
+        return E[103].fstring(T)
 
 
 @_templates.register
@@ -8360,7 +8340,7 @@ class Klass:
 
     @property
     def expected(self) -> str:
-        return E103
+        return E[103].fstring(T)
 
 
 @_templates.register
@@ -8389,8 +8369,8 @@ def function_3({CHECK}param1, {CHECK}param2, {CROSS}param3) -> {CHECK}None:
     :param None: {CROSS}
     \"\"\"
 
-{E103}
-{E203.format(directive="enable", option="unknown")}
+{E[103].fstring(T)}
+{E[203].fstring(T).format(directive="enable", option="unknown")}
 
 """
 
@@ -8421,7 +8401,7 @@ class _Messages:
 
     def __init__() -> ✓None:
 
-{E114}
+{E[114].fstring(T)}
 
 module/file.py:6 in _Messages
 -----------------------------
@@ -8431,6 +8411,6 @@ def fromcode({CHECK}ref) -> {CROSS}Message:
     :return: {CROSS}
     \"\"\"
 
-{E105}
+{E[105].fstring(T)}
 
 """