+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - option to check methods belonging to protected classes
 
 ### Removed
+- error constants in `docsig.messages`
 - hint constants in `docsig.messages`
 
 [0.40.0](https://github.com/jshwi/docsig/releases/tag/v0.40.0) - 2023-11-25